+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Allow `:worker` to be a Proc, not just a lambda
+
 ## [0.1.2] - 2025-10-29
 
 - Fix `@coordinator` RBS type to allow nil